+++ tests/svhn_neg_loader.py
@@ -4,44 +4,12 @@ import region_proposal as rp
 import file_io
 import cv2
 import os
+import annotation as ann
 
 import matplotlib.pyplot as plt
 
 N_IMAGES = 2
 DIR = '../../datasets/svhn/train'
-
-
-def get_bbs(annotation):
-    
-    bbs = []
-    labels = []
-    
-    for box in annotation["boxes"]:
-        x1 = int(box["left"])
-        y1 = int(box["top"])
-        w = int(box["width"])
-        h = int(box["height"])
-
-        bb = (y1, y1+h, x1, x1+w)
-        label = int(box["label"])
-        
-        bbs.append(bb)
-        labels.append(label)
-    return bbs, labels
-        
-def get_annotation(filename, annotation_file):
-    annotations = file_io.FileJson().read(annotation_file)
-    
-    _, filename_ = os.path.split(filename)
-    index = int(filename_[:filename_.rfind(".")])
-    annotation = annotations[index-1]
-    
-    if annotation["filename"] == filename_:
-        bbs, labels = get_bbs(annotation)
-        return bbs, labels
-        
-    else:
-        print "Annotation file should be sorted!!!!"
 
 # 1. file 을 load
 files = file_io.list_files(directory='../../datasets/svhn/train', n_files_to_sample=N_IMAGES, random_order=False)
@@ -53,7 +21,7 @@ for image_file in files:
     image = cv2.imread(image_file)
     candidates = detector.detect(image, True)
     
-    gts, labels = get_annotation(image_file, annotation_file)
+    gts, labels = ann.get_annotation(image_file, annotation_file)
     
     for ground_truth in gts:
         y1, y2, x1, x2 = ground_truth
